// frontend-admin/src/api/generated/model/expiringLicenseInfo.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export interface ExpiringLicenseInfo {
  daysRemaining?: number;
  /** @nullable */
  licenseKey?: string | null;
  licenseSaleId?: string;
  /** @nullable */
  tenantEmail?: string | null;
  tenantId?: string;
  /** @nullable */
  tenantName?: string | null;
  /** @nullable */
  tenantSlug?: string | null;
  validUntilUtc?: string;
}